Moved exception classes to `Exception` namespace
